refactor: Unify crossruntime ContentPresenter
