+++ src/Uno.UI/UI/Xaml/Controls/ContentPresenter/ContentPresenter.cs
@@ -827,6 +827,8 @@ namespace Microsoft.UI.Xaml.Controls
 			TryDetachNativeElement();
 		}
 
+		partial void ClearBorder();
+
 		private bool ResetDataContextOnFirstLoad()
 		{
 			if (!_firstLoadResetDone)
@@ -999,16 +1001,10 @@ namespace Microsoft.UI.Xaml.Controls
 
 		partial void RegisterContentTemplateRoot();
 
-		protected override void OnBackgroundChanged(DependencyPropertyChangedEventArgs e)
-		{
-			// Don't call base, the UpdateBorder() method handles drawing the background.
-			// base.OnBackgroundChanged(e);
-
-			UpdateBorder();
-#if __WASM__
-			SetAndObserveBackgroundBrush(e.OldValue as Brush, e.NewValue as Brush);
-#endif
-		}
+		/// <remarks>
+		/// Don't call base, the UpdateBorder() method handles drawing the background.
+		/// </remarks>
+		protected override void OnBackgroundChanged(DependencyPropertyChangedEventArgs e) => UpdateBorder();
 
 		internal override void UpdateThemeBindings(ResourceUpdateReason updateReason)
 		{
